resolved conflicts on filter
